Move command check to its own package
