+++ pkg/compliance/check.go
@@ -6,8 +6,14 @@
 package compliance
 
 import (
+	"time"
+
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/DataDog/datadog-agent/pkg/compliance/event"
+)
+
+const (
+	DefaultTimeout = 30 * time.Second
 )
 
 // Check is the interface for compliance checks
