+++ pkg/compliance/resources/command/command_check_test.go
@@ -6,7 +6,7 @@
 //go:build !windows
 // +build !windows
 
-package checks
+package command
 
 import (
 	"context"
